Consolidate tag helper tests into TagHelperTest.cs
* Removed `CommentContentTagHelperTest.cs`, `PostToFacebookTagHelperTest.cs`, and `PostToLinkedInTagHelperTest.cs`.
* Moved all test cases from the removed files into `TagHelperTest.cs`.
